+++ lib/plivo/base.rb
@@ -6,9 +6,9 @@ module Plivo
   module Base
     PLIVO_API_URL = 'https://api.plivo.com'.freeze
 
-    API_VOICE = 'https://voice.plivo.com'.freeze
-    API_VOICE_FALLBACK_1 = 'https://voice-usw1.plivo.com'.freeze
-    API_VOICE_FALLBACK_2 = 'https://voice-use1.plivo.com'.freeze
+    API_VOICE = 'https://api.plivo.com'.freeze
+    API_VOICE_FALLBACK_1 = 'https://api.plivo.com'.freeze
+    API_VOICE_FALLBACK_2 = 'https://api.plivo.com'.freeze
 
     CALLINSIGHTS_API_URL = 'https://stats.plivo.com'.freeze
     PHLO_API_URL = 'https://phlorunner.plivo.com'.freeze