+++ spec/resource_multipartycalls_spec.rb
@@ -77,6 +77,7 @@ describe 'MultiPartyCalls test' do
                          'agent_hold_music_method': 'GET',
                          'wait_music_method': 'GET',
                          'max_participants': 10,
+                         'record_min_member_count': 1,
                          'max_duration': 14400,
                          'ring_timeout': 45,
                          'dial_music': 'Real',
@@ -114,13 +115,13 @@ describe 'MultiPartyCalls test' do
     contents = File.read(Dir.pwd + '/spec/mocks/multiPartyCallsStartRecordingResponse.json')
     mock(202, JSON.parse(contents))
     expect(JSON.parse(to_json_MPC_start_record(@api.multipartycalls.start_recording({friendly_name: 'Voice', file_format: 'wav',
-                                                                                     status_callback_url: 'https://plivo.com/status',
-                                                                                     status_callback_method: 'POST'})))).to eql(JSON.parse(contents))
+                                                                                     recording_callback_url: 'https://plivo.com/status',
+                                                                                     recording_callback_method: 'POST'})))).to eql(JSON.parse(contents))
     compare_requests(uri: '/v1/Account/MAXXXXXXXXXXXXXXXXXX/MultiPartyCall/name_Voice/Record/',
                      method: 'POST',
                      data: {'file_format': 'wav',
-                            'status_callback_url': 'https://plivo.com/status',
-                            'status_callback_method': 'POST'
+                            'recording_callback_url': 'https://plivo.com/status',
+                            'recording_callback_method': 'POST'
                      })
   end
 
@@ -191,13 +192,13 @@ describe 'MultiPartyCalls test' do
     contents = File.read(Dir.pwd + '/spec/mocks/multiPartyCallsStartParticipantRecordingResponse.json')
     mock(200, JSON.parse(contents))
     expect(JSON.parse(to_json_MPC_start_record(@api.multipartycalls.start_participant_recording({member_id: 10, friendly_name: 'Voice', file_format: 'wav',
-                                                                                     status_callback_url: 'https://plivo.com/status',
-                                                                                     status_callback_method: 'POST'})))).to eql(JSON.parse(contents))
+                                                                                                 recording_callback_url: 'https://plivo.com/status',
+                                                                                                 recording_callback_method: 'POST'})))).to eql(JSON.parse(contents))
     compare_requests(uri: '/v1/Account/MAXXXXXXXXXXXXXXXXXX/MultiPartyCall/name_Voice/Participant/10/Record/',
                      method: 'POST',
                      data: {'file_format': 'wav',
-                            'status_callback_url': 'https://plivo.com/status',
-                            'status_callback_method': 'POST'
+                            'recording_callback_url': 'https://plivo.com/status',
+                            'recording_callback_method': 'POST'
                      })
   end